Judgementを基底enumにし、JudgementStatusにて状態管理を行うよう修正.
